修复NSHTTPCookieStorage+GGCookie 的deleteCookie方法，在Storage删除cookie后，再加上UIWebView执行js删除cookie

+++ GGCategories/GGCategories/UIView+GGBorder.m
@@ -37,7 +37,7 @@
             [self.layer insertSublayer:line atIndex:0];
         }
         UIRectEdge edge = [self gg_updateLine:line toEdge:lineEdges];
-        line.name = [NSString stringWithFormat:@"ggline%lu", (NSUInteger)edge];
+        line.name = [NSString stringWithFormat:@"ggline%lu", (unsigned long)edge];
         lineEdges -= edge;
     }
 }
